Remove RequestErrorAssert trait
JIRA: https://wavesplatform.atlassian.net/browse/NODE-580
