+++ it/src/test/scala/com/wavesplatform/it/sync/MassTransferTransactionSuite.scala
@@ -69,7 +69,7 @@ class MassTransferTransactionSuite extends BaseTransactionSuite with CancelAfter
     val (balance2, eff2) = notMiner.accountBalances(secondAddress)
     val transfers = List(Transfer(secondAddress, balance1 / 2), Transfer(thirdAddress, balance1 / 2))
 
-    assertBadRequest2(sender.massTransfer(firstAddress, transfers, calcFee(transfers.size)))
+    assertBadRequest(sender.massTransfer(firstAddress, transfers, calcFee(transfers.size)))
 
     nodes.waitForHeightAraise()
     notMiner.assertBalances(firstAddress, balance1, eff1)
@@ -82,7 +82,7 @@ class MassTransferTransactionSuite extends BaseTransactionSuite with CancelAfter
     val (balance2, eff2) = notMiner.accountBalances(secondAddress)
     val transfers = List(Transfer(secondAddress, transferAmount))
 
-    assertBadRequest2(sender.massTransfer(firstAddress, transfers, transferFee))
+    assertBadRequest(sender.massTransfer(firstAddress, transfers, transferFee))
     nodes.waitForHeightAraise()
     notMiner.assertBalances(firstAddress, balance1, eff1)
     notMiner.assertBalances(secondAddress, balance2, eff2)
@@ -96,7 +96,7 @@ class MassTransferTransactionSuite extends BaseTransactionSuite with CancelAfter
     val leaseTxId = sender.lease(firstAddress, secondAddress, leasingAmount, leasingFee).id
     nodes.waitForHeightAraiseAndTxPresent(leaseTxId)
 
-    assertBadRequest2(sender.massTransfer(firstAddress, transfers, transferFee))
+    assertBadRequest(sender.massTransfer(firstAddress, transfers, transferFee))
     nodes.waitForHeightAraise()
     notMiner.assertBalances(firstAddress, balance1 - leasingFee, eff1 - leasingAmount - leasingFee)
     notMiner.assertBalances(secondAddress, balance2, eff2 + leasingAmount)
@@ -119,7 +119,7 @@ class MassTransferTransactionSuite extends BaseTransactionSuite with CancelAfter
     for (tx <- invalidTransfers) {
       val id = tx.id()
       val req = createSignedMassTransferRequest(tx)
-      assertBadRequest2(sender.signedMassTransfer(req))
+      assertBadRequest(sender.signedMassTransfer(req))
       nodes.foreach(_.ensureTxDoesntExist(id.base58))
     }
   }
@@ -156,7 +156,7 @@ class MassTransferTransactionSuite extends BaseTransactionSuite with CancelAfter
     def id(obj: JsObject) = obj.value("id").as[String]
 
     val noProof = signedMassTransfer() - "proofs"
-    assertBadRequest2(sender.postJson("/transactions/broadcast", noProof))
+    assertBadRequest(sender.postJson("/transactions/broadcast", noProof))
     nodes.foreach(_.ensureTxDoesntExist(id(noProof)))
 
     val withProof = signedMassTransfer()
