+++ it/src/test/scala/com/wavesplatform/it/sync/transactions/AliasTransactionSuite.scala
@@ -39,7 +39,7 @@ class AliasTransactionSuite extends BaseTransactionSuite with TableDrivenPropert
     val aliasFee = calcAliasFee(firstAddress, alias)
     notMiner.assertBalances(firstAddress, balance1 - aliasFee, eff1 - aliasFee)
 
-    assertBadRequest2(sender.createAlias(firstAddress, alias, transferFee))
+    assertBadRequest(sender.createAlias(firstAddress, alias, transferFee))
     notMiner.assertBalances(firstAddress, balance1 - aliasFee, eff1 - aliasFee)
   }
 
@@ -48,7 +48,7 @@ class AliasTransactionSuite extends BaseTransactionSuite with TableDrivenPropert
 
     val (balance1, eff1) = notMiner.accountBalances(firstAddress)
     val aliasFee = calcAliasFee(firstAddress, alias)
-    assertBadRequestAndMessage2(sender.createAlias(secondAddress, alias, transferFee), "already in the state")
+    assertBadRequestAndMessage(sender.createAlias(secondAddress, alias, transferFee), "already in the state")
     notMiner.assertBalances(firstAddress, balance1 - aliasFee, eff1 - aliasFee)
   }
 
@@ -96,7 +96,7 @@ class AliasTransactionSuite extends BaseTransactionSuite with TableDrivenPropert
 
   forAll(invalid_aliases_names) { (alias: String, message: String) =>
     test(s"Not able to create alias named $alias") {
-      assertBadRequestAndMessage2(sender.createAlias(secondAddress, alias, transferFee), message)
+      assertBadRequestAndMessage(sender.createAlias(secondAddress, alias, transferFee), message)
     }
   }
 
@@ -122,7 +122,7 @@ class AliasTransactionSuite extends BaseTransactionSuite with TableDrivenPropert
    //previous test should not be commented to run this one
   test("Not able to create aliase when insufficient funds") {
     val alias = randomAlias()
-    assertBadRequestAndMessage2(sender.createAlias(firstAddress, alias, transferFee),
+    assertBadRequestAndMessage(sender.createAlias(firstAddress, alias, transferFee),
         "State check failed. Reason: negative effective balance")
   }
